+++ openslides/core/access_permissions.py
@@ -1,5 +1,4 @@
 from ..utils.access_permissions import BaseAccessPermissions
-from ..utils.auth import GROUP_ADMIN_PK, async_in_some_groups
 
 
 class ProjectorAccessPermissions(BaseAccessPermissions):
@@ -52,9 +51,4 @@ class HistoryAccessPermissions(BaseAccessPermissions):
     Access permissions container for the Histroy.
     """
 
-    async def async_check_permissions(self, user_id: int) -> bool:
-        """
-        Returns True if the user is in admin group and has read access to
-        model instances.
-        """
-        return await async_in_some_groups(user_id, [GROUP_ADMIN_PK])
+    base_permission = "core.can_see_history"