+++ CHANGELOG.rst
@@ -21,7 +21,7 @@ Core:
  - Add a change-id system to get only new elements [#3938].
  - Switch from Yarn back to npm [#3964].
  - Added password reset link (password reset via email) [#3914, #4199].
- - Added global history mode [#3977, #4141].
+ - Added global history mode [#3977, #4141, #4369, #4373].
  - Projector refactoring [4119, #4130].
  - Fixed logo configuration if logo file is deleted [#4374].